requirements.txt
pytest
pytest-django
django
djangorestframework